server/requirements.txt
google-adk
google-genai
fastapi
uvicorn[standard]
python-dotenv
asyncpg
